app/pyproject.toml
[tool.poetry]
name = "apirest"
version = "0.1.0"
description = ""
authors = ["RAJerez <rodrigoagustin23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
black = "^24.3.0"
cryptography = "^42.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
